+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Output nested signatures when a file has been signed with multiple hash algorithms
 * Renamed output property `HashAlgorithmName` to just `HashAlgorithm`
 * Added the following cmdlets
+  * [Add-OpenAuthenticodeSignature](./docs/en-US/Add-OpenAuthenticodeSignature.md) - adds a signature to an existing signature set
   * [Clear-OpenAuthenticodeSignature](./docs/en-US/Clear-OpenAuthenticodeSignature.md) - removes any signatures on a file
 
 ## v0.1.0 - 2023-03-17